requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-mock==3.14.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
urllib3==2.2.2
